# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Piece of Cake Junk - API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AdminCredentials {
  password: string;
}

export interface AuthStatus {
  isAdmin: boolean;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  unit: string;
  /** @nullable */
  iconName?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface ServiceInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: number;
  unit: string;
  iconName?: string;
}

export interface ServicePatch {
  name?: string;
  description?: string;
  price?: number;
  unit?: string;
  iconName?: string;
  isActive?: boolean;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type BookingPaymentStatus = typeof BookingPaymentStatus[keyof typeof BookingPaymentStatus];


export const BookingPaymentStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
} as const;

export interface Booking {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  city?: string;
  postalCode?: string;
  serviceDate: string;
  serviceTime: string;
  /** @nullable */
  serviceId: number | null;
  /** @nullable */
  serviceName?: string | null;
  status: BookingStatus;
  paymentStatus: BookingPaymentStatus;
  /** @nullable */
  totalAmount: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface BookingInput {
  /** @minLength 1 */
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  address: string;
  city?: string;
  postalCode?: string;
  serviceDate: string;
  serviceTime: string;
  serviceId?: number;
  notes?: string;
  subscribeToNewsletter?: boolean;
}

export type BookingPatchStatus = typeof BookingPatchStatus[keyof typeof BookingPatchStatus];


export const BookingPatchStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type BookingPatchPaymentStatus = typeof BookingPatchPaymentStatus[keyof typeof BookingPatchPaymentStatus];


export const BookingPatchPaymentStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
} as const;

export interface BookingPatch {
  status?: BookingPatchStatus;
  paymentStatus?: BookingPatchPaymentStatus;
  totalAmount?: number;
  notes?: string;
}

export interface Subscriber {
  id: number;
  email: string;
  /** @nullable */
  name: string | null;
  createdAt: string;
}

export interface SubscriberInput {
  email: string;
  name?: string;
}

export interface EmailCampaign {
  id: number;
  subject: string;
  body: string;
  recipientCount: number;
  sentAt: string;
}

export interface EmailCampaignInput {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  body: string;
}

export interface DashboardStats {
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  cancelledBookings: number;
  unpaidBookings: number;
  paidBookings: number;
  totalSubscribers: number;
  totalRevenue: number;
  pendingRevenue: number;
  bookingsToday: number;
  bookingsThisWeek: number;
}

export type ListBookingsParams = {
status?: ListBookingsStatus;
paymentStatus?: ListBookingsPaymentStatus;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type ListBookingsPaymentStatus = typeof ListBookingsPaymentStatus[keyof typeof ListBookingsPaymentStatus];


export const ListBookingsPaymentStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
} as const;